Bug #10769
Add a new CLI script, deployment-timeout.cli, to set the timout of the
application deployment of Wildfly. The timeout value is expressed in seconds
and it is defined by the new configuration property SERVER_STARTING_TIMEOUT
in the config.properties configuration file. If Silverpeas starting is too
long, with the default timeout of 5mn, the Silverpeas deployment can then
fail for timeout error. By setting this property with a greater value,
Silverpeas can take the time it needs to initialize all of its resources.

Aside that, in order to avoid the start/debug task of the installer to fail
when Silverpeas takes a long time to start, the configuration property
SERVER_STARTING_TIMEOUT is also used as timeout for JBoss to respond to the
tasks.