Update the Groovy script about the initialization of the JCR provided by Apache Jackrabbit Oak. The script takes care of threes contexts: - an Apache Jackrabbit 2 JCR already exists: it does a backup of it and   modifies the configuration files to be ready for a migration to Apache   Jackrabbit Oak with the dedicated tool provided by Oak. It creates the   JCR home directory for Oak. The index definitions for Silverpeas   should be created during the migration process by using our own   distribution of the Apache Jackrabbit Oak migration tool. - First installation: the JCR home directory and the Apache Jackrabbit   Oak is created and initialized with index definitions for Silverpeas. - No fresh installation and a non Apache Jackrabbir 2 JCR already   exists: it assumes the JCR is managed by Oak. Browse the index   definitions in the silverpeas-oak-index.properties and for each of   them, if the definition isn't already created, creates it.
Whatever the context, the script assumes by default the content storage
of the Apache Jackrabbit Oak JCR is on the filesystem (storage = segment
in the silverpeas-oak.properties). For another storage and for an
upgrade from an Apache Jackrabbit 2 JCR or a first installation, don't
forget to update the property storage in
SILVERPEAS_HOME/configuration/silverpeas/resource/silverpeas-oak.properties.

The index definitions to create are all declared in the properties file
SILVERPEAS_HOME/configuration/silverpeas/resource/silverpeas-oak-index.properties.